backend/pyproject.toml
[project]
name = "explainmybody"
version = "0.1.0"
description = "인바디 분석 및 건강 관리 플랫폼 (FastAPI 백엔드 + 프론트엔드)"
requires-python = ">=3.11,<3.12"
dependencies = [
    # Core Framework
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.29,<1.0",
    "python-dotenv>=1.0,<2.0",
    # Database
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "alembic>=1.13,<2.0",
    # Data Validation
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "email-validator>=2.1,<3.0",
    # OCR (PaddleOCR)
    "paddlepaddle==3.2.2",
    "paddleocr==3.3.2",
    "pillow>=10.0,<13.0",
    "pypdfium2>=4.30,<6.0",
    # Body Type Analysis
    "numpy>=1.26,<3.0",
    # Utilities
    "httpx>=0.27,<1.0",
    "python-multipart>=0.0.9",
    "pgvector>=0.4.2",
    "langgraph>=1.0.7",
    "openai>=2.16.0",
    "anthropic>=0.77.0",
    "opencv-python-headless>=4.13.0.92",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
]
